outro/backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
scikit-learn
pandas
joblib
python-multipart
